requirements.txt
requests==2.27.1
environs==9.4.0
google-api-python-client==2.36.0
oauth2client==4.1.3
rollbar==0.16.2
python-dateutil==2.8.2
pytz==2021.3
aiohttp==3.8.1
